Revert "feat(sdf): accept object requests on web socket & stream responses"
This reverts commit 455a86f7a8a09bb0a1dd80963a2c91303aa2792d.
